pyproject.toml
[tool.poetry]
name = "primex-web"
version = "0.1.0"
description = ""
authors = ["user <panpanzov.nur@bk.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
djangorestframework = "^3.14.0"
pillow = "^10.0.1"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
